third_party/WebKit/LayoutTests/imported/web-platform-tests/html/browsers/the-window-object/window-indexed-properties-expected.txt

This is a testharness.js-based test.
PASS Indexed properties of the window object (non-strict mode) 
PASS Indexed properties of the window object (non-strict mode) 1 
FAIL Indexed properties of the window object (non-strict mode) 2 assert_equals: expected (undefined) undefined but got (string) "foo"
PASS Indexed properties of the window object (non-strict mode) 3 
Harness: the test ran to completion.
